feat(db) Add a Triton setup field to device table
We would like to be able to track if a system is setup in Triton or not.
This migration adds a field to the device table that, when flipped,
let's us know when a system was added to Triton.

This field will primarily be used by CN setup scripts that poll Conch to
find hosts that are ready to be added to Triton.

This gets us a step closer in getting a "first seen" to "production"
metric, as a side effect.