chore: remove deploy directory
This has bit-rotted as we've maintained private versions of these
playbooks. We kept this directory for public consumption, but we
have made changes that make these no longer useful.

If we ever want to provide our playbooks (and after they've stable), we
can port them back to this repository with adequate documentation.