fix: treat email addresses as if they were case insensitive for lookups
this closes issue #141.

This change contains a database migration which creates a backup of the
user_account table before deleting duplicate entries.

----

(Instead of using lower(), we could change user_account.email from type
'text' to type 'citext', and then adding this to the schema (but it
requires superuser privileges to execute, so migrating might be
awkward):

    CREATE EXTENSION IF NOT EXISTS citext;