Fix the mismatched hardware in "passing report" we use for tests
The rack layout specifies that this slot should hold a "Test Compute" hardware (2-ssds-1-cpu),
but the device report then specified the hardware as being of type "Test Storage"
(65-ssds-2-cpu) which is not correct.

We could potentially check for this in
Conch::DB::ResultSet::DeviceLocation::assign_device_location when creating the device, but this
is probably the wrong place to do this check and we should be doing this check in a validation
instead (which we do -- albeit not thoroughly enough -- in
Conch::Validation::DeviceProductName).