add default permission checks on /workspace/:workspace_id/* endpoints
Some of these checks may be redundant with queries done in the endpoints
themselves, but this will get cleared up over time.
For now, the initial assumption is that
- HEAD, GET endpoints require 'ro'
- POST, PUT endpoints require 'rw'
- DELETE endpoints require 'admin'

We will only return 404 for workspace-not-found if the user has 'admin'
permissions on the GLOBAL workspace.

Also add a helper to abstract out the common db query.