fix creation, re-use of validation_state records
On plan execution, we actually want to search for the latest validation_state entry, regardless
of status, to consider reusing it. If the latest state entry is marked complete, then we create
a new one; otherwise, we pick up where we left off.

This also means my fix in #483 was not quite right. :/

This fix is intended for v2.19; in v2.21 we will always create new validation_state records, as
part of changes for #459.

(cherry picked from commit d4e5b81283f3f09791fe9d7000527db24679a01d)