expose schemas via their own endpoint
The API is entirely based around a series of JSON-schemas for both input
and output validation. If we expose that we can promote better utilities
for interacting with the Conch API. For an example this commit updates
the `misc/validate` script to pull from the production version of this
endpoint.