fix catching unexpected exceptions in validations
Mojo::Exception contains no additional useful information that the bare exception did not
already contain, and creates its own stack trace which is not relevant to where the actual
exception came from if it was outside of conch code. Instead we simply parse that information
out of the exception string itself, if available.