make device a required attribute, derive hardware_product from device or device_location
Since a device is always passed to validations at runtime, the attribute should be required and
a generic device should always be provided in tests.  This now also means that a
hardware_product is always created also (but not necessarily a hardware_product_profile)

We cannot make the hardware_product attribute not-providable-via-constructor until after it has
been converted to DBIC, as Conch::Model::HardwareProduct::lookup's query requires the existence
of a hardware_product_profile.