populate validation_state rows from old device_validate backup
This is to be run ahead of bin/conch thin_device_reports, which backfills results from archived
device_validate rows into validation_state.  Only overall error/fail/pass results are stored,
not the reasons for the results (which can be determined by looking at the individual
device_validate.validation json blobs).  All results for a given device_report_id are combined
together into a single validation_state entry, and are linked to a new validation_plan_id
created expressly for this script.

The input CSV file is a file prepared offline from a device_validate backup and shall be made
available on the target system(s) for running after deployment.

closes #660.