update_validation_plans script
This is suitable to be run immediately after a new code deployment (even
automatically, via ansible), before device reports start being accepted again
-- it goes through all active validation plans and updates them (via
deactivation and creating new entries with the same name) to reference the
latest version of all validations.  Therefore, validation code can be updated
in a deployment and no manual database work is necessary to update the plans.

closes #546.