validation_status tidy: drop "processing" from state, make "completed" not null
"processing" has not been used as a state since ~Sept 2018, before which it
was used as the initial state to put a validation_state record when it was
created, and then (in theory) it was updated to its final value of
error/fail/pass when report processing was complete.  Some records ended up
stuck in the initial state due to a bug that caused new records to be created
instead.

The remaining records stuck in limbo can be found and put into a more
reasonable state.