accomodate atomically swapping device serials
we set up some constraints to be deferrable to allow this.

previously:

let foreign key constraints be deferrable

Normal behaviour does not change, but inside a transaction we can now
do "set constraints all deferred", which lets us violate constraints
as long as we set everything right again by the time we "commit".