these errors should return 409 CONFLICT
That is, the state of the current data prevents the requested operation from
occuring, but it is possible that changes could be made to allow this request
in the future (missing dependent data can be added, a duplicate entry removed,
etc).

In contrast, 400 errors indicate that the request can never possibly
succeed, no matter what is changed in the system -- malformed data was sent, a
reserved name was requested, etc.