fix: log the same response that we are returning
This discrepancy occurred in places where $c->status(..) or $c->rendered(..)
was called more than once in a dispatch cycle. The after_dispatch hook is only
called once, for the first call, but what is returned in the response reflects
the last call.

- refactor the txn_wrapper helper to not actually render a response, but just
use the return value to signal the outcome; the inclusion of the first line of
the exception in the response has moved to the status helper.  Caught
exceptions are now included in the dispatch log.

- also added a check in the after_render hook to ensure this does not happen again.