simplify GET /device/*/validation_state to only return one result
Previously, it returned the latest (with matching status(es)) validation_state
per plan, but we have never in practice been switching validation_plans on a
single device, so practically speaking we never returned more than one result,
and all the existing clients only display the last result anyway.

Note that HTTP 404 will now be returned if there is no suitable validation_state
to be returned, as opposed to a previous response of 200 + "[]".