development aid: print unexpected exception that occurs after rendering a response
Exceptions are caught by Mojolicious.pm: $self->hook(around_dispatch => \&_exception);
which triggers $c->helpers->reply->exception($@)
which renders a(nother) response in development mode.

We already catch attempts to render twice, but the exception itself is lost so
the developer is left wondering what happened. We have the exception in the
stash, so just dump it here!